% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlots.R
\name{createPlot1}
\alias{createPlot1}
\title{Create interactive version of plot at page 4}
\usage{
createPlot1(width = NULL, height = NULL)
}
\arguments{
\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object
}
\description{
Create interactive version of plot at page 4
}
