% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathReport.R
\name{getPathReport}
\alias{getPathReport}
\title{get path of report available in the package}
\usage{
getPathReport()
}
\value{
string with path of report
}
\description{
get path of report available in the package
}
\author{
Laure Cougnaud and Kirsten Van Hoorde
}
