<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{printer}
\alias{printer}
\title{Print for debugging}
\usage{
printer(x)
}
\arguments{
\item{x}{R object that will be printed}
}
\value{
NULL, print output in the console
}
\description{
Print for debugging
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{printer}
\alias{printer}
\title{Print for debugging}
\usage{
printer(x)
}
\arguments{
\item{x}{R object that will be printed}
}
\value{
NULL, print output in the console
}
\description{
Print for debugging
}
>>>>>>> d8a27c64b779ab893b0c45bdc8196f8716aee9fb
