% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{readShapeData}
\alias{readShapeData}
\title{Read shape data from zipped file}
\usage{
readShapeData(zipFile, id = NULL)
}
\arguments{
\item{zipFile}{the pathname of the zip file}

\item{id}{character string, the name of a column in the shapefile .dbf 
containing the ID values to be used; default value is NULL}
}
\value{
a SpatialPolygonsDataFrame object, with polygons and data as provided
in the zipFile
}
\description{
Read shape data from zipped file
}
